+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## [major.minor.patch] - DD-MM-YYYY
 
-## Breaking Changes
+### Breaking Changes
 
 - Deleted `name_filter` from get multiple datasets endpoint [PR#91](https://github.com/quality-match/hari-client/pull/91)
 - removed `lidar_sensor_pose` from `PointCloudMetadata` model, because it's not unsupported in the HARI backend [PR#90](https://github.com/quality-match/hari-client/pull/90)
@@ -11,6 +11,7 @@
 - Add count annotation runs, pipelines, AINT learning data, ML annotation models, AI annotation runs methods [PR#91](https://github.com/quality-match/hari-client/pull/91)
 - Add `ilike` query operator [PR#91](https://github.com/quality-match/hari-client/pull/91)
 - Add method to download media files from dataset [PR#94](https://github.com/quality-match/hari-client/pull/94)
+- Add `CompositeLidarViewerVisualisationConfigParameters` model as possible visualisation configuration [PR#90](https://github.com/quality-match/hari-client/pull/90)
 
 #### Consistency in partially failed uploads